requirements.txt
# ============================================================
# Multi-Modal Medical AI Assistant - Requirements
# ============================================================

# --- Machine Learning & Deep Learning ---
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
torchxrayvision>=1.2.4
transformers>=4.36.0
sentence-transformers>=2.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
accelerate>=0.27.0

# --- RAG Frameworks ---
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
llama-index>=0.10.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

# --- Vector Databases ---
faiss-cpu>=1.7.4
chromadb>=0.4.22
rank-bm25>=0.2.2

# --- OCR ---
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
PyMuPDF>=1.23.0

# --- Medical NLP ---
spacy>=3.7.0
# Optional on Python 3.11 when SciSpaCy biomedical NER is needed.
# The codebase already falls back gracefully without it.
# scispacy>=0.5.3
nltk>=3.8.1
negspacy>=1.0.3

# --- Image Processing ---
opencv-python-headless>=4.8.0
albumentations>=1.3.1
pydicom>=2.4.0
SimpleITK>=2.3.0
onnxruntime>=1.17.0

# --- Data Processing ---
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tqdm>=4.66.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
aiofiles>=23.2.1
pydantic>=2.5.0
pymongo[srv]>=4.8.0

# --- Explainability ---
shap>=0.43.0
lime>=0.2.0.1
grad-cam>=1.4.8

# --- Backend API ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# --- Frontend ---
streamlit>=1.30.0
streamlit-extras>=0.3.6
plotly>=5.17.0

# --- Database & Storage ---
sqlalchemy>=2.0.0
alembic>=1.13.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# --- Utilities ---
loguru>=0.7.2
rich>=13.7.0
click>=8.1.7
